Introduce new packages oe for the main and the other to the differents models, in the main i create the main class Biblioteca. in the model packages i create the class DatoLibro and the class DatosAlumno

+++ Pruebas-ra4-Exam/src/model/DatoLibro.java
@@ -15,13 +15,15 @@ public class DatoLibro {
     
     private String Title;
     private String Autor;
-    private Integer PoD;
+    private long PoD;
 
-    public DatoLibro(String Title, String Autor, Integer PoD) {
+    public DatoLibro(String Title, String Autor, long PoD) {
         this.Title = Title;
         this.Autor = Autor;
         this.PoD = PoD;
     }
+
+  
 
     public String getTitle() {
         return Title;
@@ -39,14 +41,17 @@ public class DatoLibro {
         this.Autor = Autor;
     }
 
-    public Integer getPoD() {
+    public long getPoD() {
         return PoD;
     }
 
-    public void setPoD(Integer PoD) {
+    public void setPoD(long PoD) {
         this.PoD = PoD;
     }
 
+   
+
+
